Calendar refactor to use hooks and store
