+++ client/src/components/ScheduleView/Calendar.css
@@ -252,7 +252,8 @@ transition: 0.5s ease-in;
   }
   
   .calendar .body .disabled {
-    color: var(--text-color-light);
+    color: black;
+    background-color: lightgray;
     pointer-events: none;
   }
